debugger/backend/requirements.txt
slowapi==0.1.9
limits==3.6.0
docker==7.1.0
urllib3==2.2.3
requests==2.32.3
chardet==5.2.0
idna==3.10
certifi==2024.8.30
fastapi==0.115.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12
uvicorn==0.32.1
starlette==0.41.3
anyio==4.6.2
pydantic==1.10.21
email-validator==2.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1
typing-extensions==4.12.2
click==8.1.7
h11==0.14.0
sniffio==1.3.1
mako==1.3.8
MarkupSafe==3.0.2
httpx==0.27.2
authlib==1.3.2
redis==5.2.1
